Refactor ticket info method in Tickets class
This commit changes the way ticket information is returned in the info method inside Tickets class. The method switches from sending a POST call with URL parameters to sending a POST call with a JSON-encoded body. This change was performed to potentially solve issues with query parameters being too long.